requirements.txt
langchain==0.0.335
langchain-openai==0.0.2.post1
streamlit==1.28.0
python-dotenv==1.0.0
openai>=1.6.1,<2.0.0
streamlit-chat==0.1.1 